+++ lib/src/widgets/internal/controls/controls.dart
@@ -1,2 +1,3 @@
 export 'flutter_deck_controls.dart';
+export 'flutter_deck_controls_listener.dart';
 export 'flutter_deck_controls_notifier.dart';